move server implementation to tracker module
